Some wires must be defined within a When Stmt
If a wire is defined within a When stmt, it is essentially just renaming
some existing wires.
If a wire is defined in a root stmt block, then you need to assign a
default stmt.
Wires defined in a root stmt block can be used as starting points for
topological sorting while other wires cannot be.